requirements.txt
azure-identity
azure-ai-projects
azure-functions
python-dotenv